# default/openapi.yaml
openapi: 3.0.1
info:
  title: api.hangulize.org
  version: 0.0.0
  license:
    name: MIT
servers:
- url: https://api.hangulize.org/v2
  description: Production server
paths:
...
  /hangulized/{lang}/{word}:
    get:
      summary: Transcribes a non-Korean word into Hangul.
      parameters:
      - $ref: '#/components/parameters/lang'
      - $ref: '#/components/parameters/word'
      responses:
        '200':
          description: A transcription result.
          content:
            text/plain:
              schema:
                type: string
              example: 글로리아
            application/json:
              schema:
                type: object
                properties:
                  lang:
                    type: string
                  word:
                    type: string
                  transcribed:
                    type: string
              example: {
                "lang": "ita",
                "word": "gloria",
                "transcribed": "글로리아"
              }
...
components:
...
  parameters:
    lang:
      name: lang
      description: Langauge ID
      in: path
      required: true
      schema:
        type: string
      example: ita
...
    word:
      name: word
      description: Non-Korean word
      in: path
      required: true
      schema:
        type: string
      example: gloria